fixed slug
